requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3
langchain==0.0.267
flet==0.27.6
pypdf2==3.0.1
python-docx==0.8.11
markdown==3.4.4
sentence-transformers==2.2.2
huggingface-hub>=0.16.4
numpy>=1.20.0
torch>=1.6.0
transformers>=4.26.0